refactor(snmp): decompose the interface god-file by query area (WS-D)
internal/protocols/snmp/interface.go was a 1196-line god-file bundling every
SNMP interface query area behind one file. Split by query area into role files
within the same package (pure move — no API, signature or logic change; verified
by an identical top-level declaration set and an identical sorted code-line
multiset against the original):

- interface.go (360) — shared IF-MIB/BRIDGE-MIB OID + status consts, InterfaceInfo
                       /MACEntry types, GetInterfaceInfo/GetAllInterfaces + walks
- mactable.go (350)  — GetMACTable and the Q-BRIDGE/BRIDGE-MIB FDB walkers/parsers
- portvlan.go (148)  — GetPortVLANs + bridge-port-to-ifIndex mapping
- parse.go (124)     — interface/duplex/MAC status + OID/TimeTicks parse helpers
- counters.go (263)  — InterfaceCounters + GetInterfaceCounters + 32/64-bit walkers

The shared OID/status/parsing consts stay in interface.go (package-scoped, used
across the new files). Largest resulting file is 360 LOC, clearing the file-size
threshold and removing interface.go as a STRICT-gate blocker for WS-Z.

Advances the Architecture Completion Plan WS-D (god-file decomposition).
